function/requirements.txt
google-cloud-secret-manager
pg8000
